+++ src/mruby_integration/audio/music.cpp
@@ -6,28 +6,6 @@
 #include "mruby_integration/struct_types.hpp"
 
 auto
-mrb_pause_music_stream(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  Music* music;
-  mrb_get_args(mrb, "d", &music, &Music_type);
-
-  PauseMusicStream(*music);
-
-  return mrb_nil_value();
-}
-
-auto
-mrb_resume_music_stream(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  Music* music;
-  mrb_get_args(mrb, "d", &music, &Music_type);
-
-  ResumeMusicStream(*music);
-
-  return mrb_nil_value();
-}
-
-auto
 mrb_set_music_volume(mrb_state* mrb, mrb_value) -> mrb_value
 {
   Music* music;
@@ -63,16 +41,6 @@ mrb_get_music_time_length(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_audio_music(mrb_state* mrb)
 {
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "pause_music_stream",
-                    mrb_pause_music_stream,
-                    MRB_ARGS_REQ(1));
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "resume_music_stream",
-                    mrb_resume_music_stream,
-                    MRB_ARGS_REQ(1));
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "set_music_volume",