+++ CHANGELOG.md
@@ -94,6 +94,8 @@
 - Removed get_music_time_played
 - Music#update is now C++ backed
 - Removed update_music_stream
+- Music#pause and Music#resume are now C++ backed
+- Removed pause_music_stream and resume_music_stream
 
 ## v0.3.14.1